Fix: Integer overflow with small step sizes.

+++ Script/Waher.Script.Graphs/Graph.cs
@@ -856,7 +856,7 @@ namespace Waher.Script.Graphs
 
 			double StepSize = Num / Den;
 			List<double> Steps = new List<double>();
-			int i = (int)Math.Ceiling(Min / StepSize);
+			double i = Math.Ceiling(Min / StepSize);
 			int j = 0;
 			double d = (i * Num) / Den;
